+++ packages/bucket-source-github/src/GitHub.js
@@ -39,7 +39,7 @@ export const GitHub = ({ owner, repo, basePath = '', auth, headers = { 'X-GitHub
           _slug: file.name,
           _filename: file.name,
           ..._extractFrontMatter(fileContents),
-          ...(omitBody === false ? _extractBody(fileContents) : { body: null }),
+          ...(omitBody === false ? _extractBody(fileContents) : { _body: null }),
         }
       });
       return await Promise.all(promises);
@@ -65,7 +65,7 @@ export const GitHub = ({ owner, repo, basePath = '', auth, headers = { 'X-GitHub
   const _extractBody = (file) => {
     const body = file.split('---')[2];
     if(body === undefined) throw new Error('Can not extract body, file may be formatted incorrectly.');
-    return { body };
+    return { _body: body };
   }
   const _extractFrontMatter = (file) => {
     // via and thanks to: https://github.com/jxson/front-matter/blob/master/index.js